+++ layers/Engine/packages/component-model/config/services.yaml
@@ -35,9 +35,6 @@ services:
     PoP\ComponentModel\DirectivePipeline\DirectivePipelineServiceInterface:
         class: \PoP\ComponentModel\DirectivePipeline\DirectivePipelineService
 
-    PoP\ComponentModel\Engine\MandatoryFieldDirectiveResolverRegistryInterface:
-        class: \PoP\ComponentModel\Engine\MandatoryFieldDirectiveResolverRegistry
-
     PoP\ComponentModel\Feedback\FeedbackEntryManagerInterface:
         class: \PoP\ComponentModel\Feedback\FeedbackEntryManager
 
@@ -119,6 +116,9 @@ services:
 
     PoP\ComponentModel\ComponentProcessors\ComponentProcessorManagerInterface:
         class: \PoP\ComponentModel\ComponentProcessors\ComponentProcessorManager
+
+    PoP\ComponentModel\Registries\MandatoryFieldDirectiveResolverRegistryInterface:
+        class: \PoP\ComponentModel\Registries\MandatoryFieldDirectiveResolverRegistry
 
     PoP\ComponentModel\Registries\TypeRegistryInterface:
         class: \PoP\ComponentModel\Registries\TypeRegistry
