+++ layers/Engine/packages/component-model/src/Container/CompilerPasses/RegisterMandatoryDirectiveServiceTagCompilerPass.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace PoP\ComponentModel\Container\CompilerPasses;
 
 use PoP\ComponentModel\Container\ServiceTags\MandatoryDirectiveServiceTagInterface;
-use PoP\ComponentModel\Engine\MandatoryFieldDirectiveResolverRegistryInterface;
+use PoP\ComponentModel\Registries\MandatoryFieldDirectiveResolverRegistryInterface;
 use PoP\Root\Container\CompilerPasses\AbstractInjectServiceIntoRegistryCompilerPass;
 
 class RegisterMandatoryDirectiveServiceTagCompilerPass extends AbstractInjectServiceIntoRegistryCompilerPass
